net/mlx5: fix CT objects limited to 65k
The dev_idx variable is responsible for holding the
index of the CT (connection tracking) object.
The variable has a type of uint16_t, this means
the largest index possible is 2^16 = 65,536.
Change the type to be uint32_t so that it can hold larger values.

Fixes: 20f8000f8285 ("net/mlx5: add support of CT between two ports")

+++ drivers/net/mlx5/mlx5_flow_dv.c
@@ -14726,7 +14726,7 @@ __flow_dv_action_ct_update(struct rte_eth_dev *dev, uint32_t idx,
 	const struct rte_flow_action_conntrack *new_prf;
 	int ret = 0;
 	uint16_t owner = (uint16_t)MLX5_ACTION_CTX_CT_GET_OWNER(idx);
-	uint16_t dev_idx;
+	uint32_t dev_idx;
 
 	if (PORT_ID(priv) != owner)
 		return rte_flow_error_set(error, EACCES,
